- Add color button backgrounds for main screen buttons - Add new start screen - Removed the selection buttons from the flash card activity - Added ads - Wired up the main screen buttons - Hide the division button - Add designated landscape layouts - Wired up for rotation to persist current screen
