+++ app/src/main/java/dwyer/com/mathflashcards/FlashCardActivity.java
@@ -1,0 +1,27 @@
+package dwyer.com.mathflashcards;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.MobileAds;
+
+public class FlashCardActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_flash_card);
+
+        MobileAds.initialize(getApplicationContext(), "ca-app-pub-0959410565463062~3421628515");
+
+        AdView mAdView = (AdView) findViewById(R.id.adView);
+        AdRequest adRequest = new AdRequest.Builder().addTestDevice("990D881316626793E9B06C55C545B512").build();
+        //AdRequest adRequest = new AdRequest.Builder().build();
+        mAdView.loadAd(adRequest);
+
+    }
+
+
+}
