+++ app/src/main/res/drawable/buttons_main_screen_red.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#fd0400ff"/>
+
+    <stroke android:width="2dp"
+        android:color="#000000"/>
+
+    <padding android:top="2dp"
+        android:right="10dp"
+        android:bottom="2dp"
+        android:left="10dp"/>
+
+    <gradient
+        android:startColor="#fdfc0101"
+        android:centerColor="#fdff9094"
+        android:endColor="#fdfc0101"
+        android:angle="90"
+        android:useLevel="true"/>
+
+    <corners android:radius="5dp"/>
+
+</shape>
