+++ app/src/main/res/drawable/buttons_submit.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#fd0400ff"/>
+
+    <stroke android:width="3dp"
+        android:color="#000000"/>
+
+    <padding android:top="2dp"
+        android:right="10dp"
+        android:bottom="2dp"
+        android:left="10dp"/>
+
+    <corners android:radius="50dp"/>
+
+</shape>
