--- app/src/main/res/drawable/function_button.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#fd4342a1"/>
-
-    <stroke android:width="3dp"
-        android:color="#000000"/>
-
-    <padding android:top="5dp"
-        android:right="5dp"
-        android:bottom="5dp"
-        android:left="5dp"/>
-
-    <corners android:radius="200dp"/>
-</shape>
