+++ app/src/main/res/menu/menu_main.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/action_rate"
+        android:title="@string/rate_us"/>
+    <item android:id="@+id/action_share"
+        android:title="@string/share" />
+</menu>
